Removing comments and debug prints, deleting empty files and moving gui class to classes folder
